moved tool to new file, fixed env usage
