+++ markdown_validator/.env.example
@@ -1,3 +1,8 @@
-MODEL='oh-2.5m7b-q51'
-OPENAI_API_BASE_URL="http://localhost:8000/v1" # LLM Studio
-OPENAI_API_KEY=local
+# Using OpenAI's API
+OPENAI_API_KEY="sk-..."
+MODEL_NAME="gpt-3.5-turbo"
+
+# Using LLM Studio's API
+# MODEL_NAME='oh-2.5m7b-q51'
+# OPENAI_API_BASE_URL="http://localhost:8000/v1"
+# OPENAI_API_KEY=local
